fixed a few more warnings (mostrly in the Clustering refactoring, my bad) fixed indentationj in TLPExport, and starting work on a pluginProgress use in it
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@2888 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-ogl/src/AbstractGlCurve.cpp
@@ -1,4 +1,4 @@
-/**
+  /**
  *
  * This file is part of Tulip (www.tulip-software.org)
  *
@@ -333,7 +333,7 @@ void AbstractGlCurve::drawCurve(std::vector<Coord> &controlPoints, const Color &
 		curveShaderProgram = curveShaderProgramNormal;
 	}
 
-	if (curveShaderProgram != NULL && controlPoints.size() <= MAX_SHADER_CONTROL_POINTS && renderMode != GL_SELECT) {
+	if (curveShaderProgram != NULL && controlPoints.size() <= (size_t)MAX_SHADER_CONTROL_POINTS && renderMode != GL_SELECT) {
 
 		GLuint *vbo = curveVertexBuffersObject[nbCurvePoints];
 		GlShaderProgram *currentActiveShader = GlShaderProgram::getCurrentActiveShader();
